fix: ADL-discoverable int128/double ops; replace long with int64_t
- numeric: move the Boost int128 double-interop shim into namespace
  boost::multiprecision so it is found via ADL. MSVC only consults ADL
  (not ordinary lookup) inside the `requires(Int x, Float g){x*g;}` probe
  in Rational's float constructor, so the global-scope shim was invisible
  there and the constructor still took the infinitely-recursive branch
  (the bentleyottmann stack overflow). ADL placement also lets doctest's
  expression decomposition find int128==double, fixing the C2440 in
  doctest's operator L() when a Boost int128 is a CHECK operand.

- Replace all uses of long / long long with int64_t across the library
  and tests. long is 32-bit on MSVC (LLP64), which overflowed e.g.
  Rational<long> squared distances in the ShapeTree tests. Collapse the
  now-duplicate BigInt operator long/long long into one operator int64_t.

Verified 46/46 on native g++ and clang++ and under a forced Boost-int128
build (g++ -U__SIZEOF_INT128__), including bentleyottmann under a 256 KB
stack (no recursion).

+++ tests/unit/polygon.cpp
@@ -95,7 +95,7 @@ TEST_CASE("Polygon intersects Polygon into points, polylines, and polygons") {
 }
 
 TEST_CASE("Polygon intersects Polygon with fractional crossings (rational result)") {
-    using Q = pgl::Rational<long long>;
+    using Q = pgl::Rational<int64_t>;
     using P = pgl::Point<int>;
     using QPoint = pgl::Point<Q>;
     using QPolygon = pgl::Polygon<QPoint>;
